chore: move changes to requestid.go
